rename to blind-player-checks
