+++ src/modules/blind-player-checks.js
